+++ src/app/register/[resource_id]/components/SplashLoaderModal.tsx
