+++ src/app/register/[resource_id]/components/index.ts
@@ -1,0 +1,2 @@
+export * from './RegisterFormScreen'
+export * from './SplashLoaderModal'
